validator/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
collision==1.2.2
pydantic==2.6.4
pydantic_core==2.16.3